refactor: remove `utxoByAddresses` from wallet provider mocks
